requirements.txt
pandas
numpy
pycaret[full]
streamlit
google-generativeai
openai
requests
